+++ previous-versions/banking/codatbanking.go
@@ -138,8 +138,8 @@ func New(opts ...SDKOption) *CodatBanking {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "3.0.0",
-			SDKVersion:        "0.22.0",
-			GenVersion:        "2.107.3",
+			SDKVersion:        "0.22.1",
+			GenVersion:        "2.108.3",
 		},
 	}
 	for _, opt := range opts {