+++ previous-versions/banking/RELEASES.md
@@ -307,3 +307,13 @@ Based on:
 - [go v0.22.0] previous-versions/banking
 ### Releases
 - [Go v0.22.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/banking/v0.22.0 - previous-versions/banking
+
+## 2023-09-12 00:16:39
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.22.1] previous-versions/banking
+### Releases
+- [Go v0.22.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/banking/v0.22.1 - previous-versions/banking